rename new_bidirectional to bidirectional for merge
